chore(types): move type converters to `types` from `util`
