+++ src/modules/types/boolToNum.js
