+++ src/modules/types/boolToStr.js
