+++ src/modules/types/numToBool.js
