+++ src/modules/types/strToBool.js
