+++ src/modules/types/strToNum.js
